requirements.txt
discord.py>=2.0
Pillow>=10.0
aiohttp>=3.0
python-dotenv>=1.0
